jetson_ads-stack-py/Dashboard/requirements.txt
pyzmq==26.2.0
pigpio==1.78
Jinja2==3.1.2
Flask==2.2.2
gpiozero==2.0.1
flask-socketio
